requirements.txt
django==3.1.7
requests==2.25.1
python-decouple==3.4